requirements.txt
python-dotenv==1.0.1
selenium==4.23.1
webdriver-manager==4.0.2
pandas==2.2.2
openpyxl==3.1.5
xlrd==2.0.1
lxml==5.2.2
beautifulsoup4==4.12.3
